requirements.txt
pandas
SQLAlchemy
psycopg2-binary
requests
python-dotenv
matplotlib
numpy
adjustText